backend/sema-api/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Translation models and processing
ctranslate2>=4.0.0,<5.0.0
sentencepiece>=0.1.99,<0.3.0
fasttext-wheel>=0.9.2,<1.0.0
huggingface_hub>=0.17.0,<1.0.0
numpy>=1.21.0,<2.0.0  # Pin to NumPy 1.x for compatibility with translation libraries
scipy>=1.7.0,<2.0.0   # Ensure compatible scipy version

# Security and rate limiting
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Utilities
pytz>=2023.3
